Former-commit-id: e8a6caf193432d17b21df151a565062b74d13ceb

+++ packages/engine/src/networking/schema/clientInputSchema.ts
@@ -1,4 +1,4 @@
-import { float32, Model, Schema, uint32, uint64, uint8 } from "superbuffer"
+import { float32, Model, Schema, uint32, uint8 } from "superbuffer"
 import { NetworkClientInputInterface, PacketNetworkClientInputInterface } from "../interfaces/WorldState";
 //import { uint8, float32, uint16, uint32 } from "../../common/types/DataTypes";
 //import { createSchema } from "../functions/createSchema";
@@ -35,7 +35,7 @@ export const inputKeyArraySchema = new Schema({
   axes2d: [inputAxis2DSchema],
   buttons: [inputKeySchema],
   viewVector: viewVectorSchema,
-  snapShotTime: uint64
+  snapShotTime: uint32
 });
 
 export class ClientInputModel {
@@ -43,7 +43,7 @@ export class ClientInputModel {
   static toBuffer(inputs: NetworkClientInputInterface): ArrayBuffer {
     const packetInputs: PacketNetworkClientInputInterface = {
       ...inputs,
-      snapShotTime: BigInt( inputs.snapShotTime )
+      snapShotTime: inputs.snapShotTime
     }
     // @ts-ignore
     return this.model.toBuffer(packetInputs);
